google-ads-bi-agent/requirements.txt
pandas==2.3.3
langchain-openai==1.1.3
langchain-core==1.2.1
pydantic==2.12.5
tabulate==0.9.0
PyYAML==6.0.3
google-ads>=23.0.0
python-dotenv==1.0.1
google-auth-oauthlib
